ref: Destructurized to separate modules for visual part and API part
