+++ packages/next/src/app/page.tsx
@@ -3,7 +3,7 @@ import '@fontsource/roboto/300.css';
 import '@fontsource/roboto/400.css';
 import '@fontsource/roboto/500.css';
 import '@fontsource/roboto/700.css';
-import { UrlTesterSection } from '@/containers/home-page/url-tester-section';
+import { UrlTesterSection } from '@/modules/frontend/containers/home-page/url-tester-section';
 
 export default function Home() {
   return (
