+++ packages/next/src/modules/frontend/containers/home-page/url-tester-section/url-input/styled.ts
@@ -1,4 +1,4 @@
-import { COLORS } from '@/theme/colors.const';
+import { COLORS } from '@/modules/frontend/theme/colors.const';
 import styled from '@emotion/styled';
 import { TextareaAutosize } from '@mui/material';
