+++ packages/next/src/modules/frontend/containers/home-page/url-tester-section/url-results/styled.ts
@@ -1,7 +1,7 @@
 import { styled } from '@mui/material';
 import { DataGrid } from '@mui/x-data-grid';
 import { darken } from '@mui/material/styles';
-import { COLORS } from '@/theme/colors.const';
+import { COLORS } from '@/modules/frontend/theme/colors.const';
 
 export const CustomDataGrid = styled(DataGrid)`
   .available {
